Backend/requirements.txt
Flask
flask-cors
tensorflow
Pillow
numpy
